requirements.txt
aiohttp
faiss-cpu
fastapi[standard]
huggingface_hub
langchain
langchain-classic
langchain-community
langchain-core
langchain-google-genai
langchain-openai
langchain-tavily
pydantic
psycopg2-binary
psycopg[binary,pool]
langgraph
langgraph-checkpoint-postgres
sqlalchemy
alembic
psycopg2-binary
pwdlib[argon2]
pyjwt
